[41] Refactor search result query logic
